+++ platform/darwin/test/MGLRendererConfigurationTests.mm
@@ -6,6 +6,7 @@ static NSString * const MGLRendererConfigurationTests_collisionBehaviorKey = @"M
 
 @interface MGLRendererConfiguration (Tests)
 - (instancetype)initWithPropertyDictionary:(nonnull NSDictionary*)bundle;
+- (mbgl::optional<std::string>)_localFontFamilyNameWithPropertyDictionary:(nonnull NSDictionary *)properties;
 @end
 
 
@@ -77,14 +78,12 @@ static NSString * const MGLRendererConfigurationTests_collisionBehaviorKey = @"M
 }
 
 - (void)testOverridingMGLCollisionBehaviorPre40 {
-
     // Dictionary = NO, NSUserDefaults = YES
     {
         [[NSUserDefaults standardUserDefaults] setObject:@(YES) forKey:MGLRendererConfigurationTests_collisionBehaviorKey];
         MGLRendererConfiguration *config = [[MGLRendererConfiguration alloc] initWithPropertyDictionary:@{MGLRendererConfigurationTests_collisionBehaviorKey:@(NO)}];
         XCTAssert(config.perSourceCollisions);
     }
-
     // Dictionary = YES, NSUserDefaults = NO
     {
         [[NSUserDefaults standardUserDefaults] setObject:@(NO) forKey:MGLRendererConfigurationTests_collisionBehaviorKey];
@@ -92,5 +91,107 @@ static NSString * const MGLRendererConfigurationTests_collisionBehaviorKey = @"M
         XCTAssertFalse(config.perSourceCollisions);
     }
 }
+
+- (void)testDefaultLocalFontFamilyName {
+    
+    MGLRendererConfiguration *config = [[MGLRendererConfiguration alloc] init];
+    std::string localFontFamilyName = config.localFontFamilyName.value();
+    
+    std::string systemFontFamilyName;
+#if TARGET_OS_IPHONE
+    systemFontFamilyName = std::string([[UIFont systemFontOfSize:0 weight:UIFontWeightRegular].familyName UTF8String]);
+#else
+    systemFontFamilyName = std::string([[NSFont systemFontOfSize:0 weight:NSFontWeightRegular].familyName UTF8String]);
+#endif
+    
+    XCTAssertEqual(localFontFamilyName, systemFontFamilyName, @"Default local font family name should match default system font");
+}
+
+- (void)testSettingMGLIdeographicFontFamilyNameWithPlistValue {
+    
+    MGLRendererConfiguration *config = [[MGLRendererConfiguration alloc] init];
+    NSDictionary *dic;
+    
+    // `MGLIdeographicFontFamilyName` set to bool value `YES`
+    {
+        dic = @{@"MGLIdeographicFontFamilyName": @(YES)};
+        std::string localFontFamilyName = ([config _localFontFamilyNameWithPropertyDictionary:dic]).value();
+        
+        std::string systemFontFamilyName;
+#if TARGET_OS_IPHONE
+        systemFontFamilyName = std::string([[UIFont systemFontOfSize:0 weight:UIFontWeightRegular].familyName UTF8String]);
+#else
+        systemFontFamilyName = std::string([[NSFont systemFontOfSize:0 weight:NSFontWeightRegular].familyName UTF8String]);
+#endif
+        XCTAssertEqual(localFontFamilyName, systemFontFamilyName, @"Local font family name should match default system font name when setting `YES`");
+    }
+    
+    // `MGLIdeographicFontFamilyName` set to bool value `NO`
+    {
+        dic = @{@"MGLIdeographicFontFamilyName": @(NO)};
+        mbgl::optional<std::string> localFontFamilyName = [config _localFontFamilyNameWithPropertyDictionary:dic];
+        XCTAssertFalse(localFontFamilyName.has_value(), @"Client rendering font should use remote font when setting `NO`");
+    }
+    
+    // `MGLIdeographicFontFamilyName` set to a valid font string value
+    {
+        dic = @{@"MGLIdeographicFontFamilyName": @"PingFang TC"};
+        std::string localFontFamilyName = ([config _localFontFamilyNameWithPropertyDictionary:dic]).value();
+        std::string targetFontFamilyName = std::string([@"PingFang TC" UTF8String]);
+        XCTAssertEqual(localFontFamilyName, targetFontFamilyName, @"Local font family name should match a custom valid font name");
+    }
+    
+    // `MGLIdeographicFontFamilyName` set to an invalid font string value
+    {
+        dic = @{@"MGLIdeographicFontFamilyName": @"test font"};
+        std::string localFontFamilyName = ([config _localFontFamilyNameWithPropertyDictionary:dic]).value();
+        
+        std::string systemFontFamilyName;
+#if TARGET_OS_IPHONE
+        systemFontFamilyName = std::string([[UIFont systemFontOfSize:0 weight:UIFontWeightRegular].familyName UTF8String]);
+#else
+        systemFontFamilyName = std::string([[NSFont systemFontOfSize:0 weight:NSFontWeightRegular].familyName UTF8String]);
+#endif
+        XCTAssertEqual(localFontFamilyName, systemFontFamilyName, @"Local font family name should match default system font name when setting an invalid font string");
+    }
+    
+    // `MGLIdeographicFontFamilyName` set to a valid font family names array value
+    {
+        dic = @{@"MGLIdeographicFontFamilyName": @[@"test font 1", @"PingFang TC", @"test font 2"]};
+        std::string localFontFamilyName = ([config _localFontFamilyNameWithPropertyDictionary:dic]).value();
+        std::string targetFontFamilyName = std::string([@"PingFang TC" UTF8String]);
+        XCTAssertEqual(localFontFamilyName, targetFontFamilyName, @"Local font family name should match a custom valid font name in a font family names array");
+    }
+    
+    // `MGLIdeographicFontFamilyName` set to an invalid font family names array value
+    {
+        dic = @{@"MGLIdeographicFontFamilyName": @[@"test font 1", @"test font 2", @"test font 3"]};
+        std::string localFontFamilyName = ([config _localFontFamilyNameWithPropertyDictionary:dic]).value();
+        
+        std::string systemFontFamilyName;
+#if TARGET_OS_IPHONE
+        systemFontFamilyName = std::string([[UIFont systemFontOfSize:0 weight:UIFontWeightRegular].familyName UTF8String]);
+#else
+        systemFontFamilyName = std::string([[NSFont systemFontOfSize:0 weight:NSFontWeightRegular].familyName UTF8String]);
+#endif
+        XCTAssertEqual(localFontFamilyName, systemFontFamilyName, @"Local font family name should match default system font name when setting an invalid font family names array");
+    }
+    
+    // `MGLIdeographicFontFamilyName` set to an invalid value type: NSDictionary, NSNumber, NSData, etc.
+    {
+        dic = @{@"MGLIdeographicFontFamilyName": [@"test font 1" dataUsingEncoding:NSUTF8StringEncoding]};
+        std::string localFontFamilyName = ([config _localFontFamilyNameWithPropertyDictionary:dic]).value();
+        
+        std::string systemFontFamilyName;
+#if TARGET_OS_IPHONE
+        systemFontFamilyName = std::string([[UIFont systemFontOfSize:0 weight:UIFontWeightRegular].familyName UTF8String]);
+#else
+        systemFontFamilyName = std::string([[NSFont systemFontOfSize:0 weight:NSFontWeightRegular].familyName UTF8String]);
+#endif
+        XCTAssertEqual(localFontFamilyName, systemFontFamilyName, @"Local font family name should match default system font name when setting an invalid value type");
+    }
+}
+
+
 
 @end