+++ src/mbgl/renderer/layers/render_raster_layer.cpp
@@ -74,8 +74,9 @@ static std::array<float, 3> spinWeights(float spin) {
 }
 
 void RenderRasterLayer::prepare(const LayerPrepareParameters& params) {
-    RenderLayer::prepare(params);
+    renderTiles = params.source->getRenderTiles();
     imageData = params.source->getImageRenderData();
+    assert(renderTiles || imageData);
 }
 
 void RenderRasterLayer::render(PaintParameters& parameters) {
@@ -153,8 +154,8 @@ void RenderRasterLayer::render(PaintParameters& parameters) {
                 },
                 bucket.drawScopeID + std::to_string(i++));
         }
-    } else {
-        for (const RenderTile& tile : renderTiles) {
+    } else if (renderTiles) {
+        for (const RenderTile& tile : *renderTiles) {
             auto* bucket_ = tile.getBucket(*baseImpl);
             if (!bucket_) {
                 continue;