+++ src/mbgl/renderer/sources/render_image_source.hpp
@@ -48,8 +48,6 @@ public:
                 bool needsRelayout,
                 const TileParameters&) final;
 
-    RenderTiles getRenderTiles() override { return {}; }
-
     const ImageSourceRenderData* getImageRenderData() const override { 
         return renderData.get();
     }
@@ -57,7 +55,7 @@ public:
     std::unordered_map<std::string, std::vector<Feature>>
     queryRenderedFeatures(const ScreenLineString& geometry,
                           const TransformState& transformState,
-                          const std::vector<const RenderLayer*>& layers,
+                          const std::unordered_map<std::string, const RenderLayer*>& layers,
                           const RenderedQueryOptions& options,
                           const mat4& projMatrix) const final;